requirements.txt
Flask>=2.0.0
cryptography>=38.0.0
PyJWT>=2.0.0
bcrypt>=3.2.0
Pillow>=9.0.0